no O(N) operations inside inits and move internal cryojax utilities to _misc.py
