+++ cryojax/simulator/_transfer_theory/transfer_theory.py
@@ -2,7 +2,8 @@ import equinox as eqx
 import jax.numpy as jnp
 from jaxtyping import Array, Complex, Float
 
-from ...jax_util import FloatLike, error_if_not_fractional
+from ..._misc import error_if_not_fractional
+from ...jax_util import FloatLike
 from ...ndimage import AbstractFourierOperator
 from .._image_config import AbstractImageConfig
 from .transfer_function import AbstractCTF
